+++ Editor/Playa/Renderer/SerializedFieldRendererUIToolkit.cs
@@ -32,10 +32,9 @@ namespace SaintsField.Editor.Playa.Renderer
                 // Debug.Log(string.Join(",", FieldWithInfo.PlayaAttributes));
                 // Debug.Log(FieldWithInfo.PlayaAttributes.Any(each => each is SaintsSerializedAttribute));
                 // Debug.Log(FieldWithInfo.PlayaAttributes.Any(each => each is SaintsSerializedAttribute));
-                // if (elementType == typeof(SaintsSerializedProperty) && label.EndsWith("__Saints Serialized__"))
-                if (label.EndsWith("__Saints Serialized__") && FieldWithInfo.PlayaAttributes.Any(each => each is SaintsSerializedAttribute))
+                if (label.EndsWith(Util.SaintsSerializedLabelSuffix) && FieldWithInfo.PlayaAttributes.Any(each => each is SaintsSerializedAttribute))
                 {
-                    label = label[..^"__Saints Serialized__".Length];
+                    label = label[..^Util.SaintsSerializedLabelSuffix.Length];
                 }
             }
             VisualElement r = UIToolkitUtils.CreateOrUpdateFieldProperty(