+++ Samples~/Scripts/SaintsEditor/Testing/GetForInterfaceTest/IMyInterface.cs
@@ -1,0 +1,7 @@
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.GetForInterfaceTest
+{
+    public interface IMyInterface
+    {
+
+    }
+}