+++ Samples~/Scripts/SaintsEditor/Testing/GetForInterfaceTest/InterMono2.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.GetForInterfaceTest
+{
+    public class InterMono2 : MonoBehaviour, IMyInterface
+    {
+    }
+}