+++ Samples~/Scripts/SaintsEditor/Testing/OnEventBindWalk/OnEventTarget.cs
@@ -1,0 +1,11 @@
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace SaintsField.Samples.Scripts.SaintsEditor.Testing.OnEventBindWalk
+{
+    public class OnEventTarget : MonoBehaviour
+    {
+        // ReSharper disable once InconsistentNaming
+        public UnityEvent _OnLocalize = new UnityEvent();
+    }
+}